+++ extensions/preprocess/html.php
@@ -19,6 +19,8 @@ function ultima_preprocess_html(&$vars) {
   // doing so we will remove all slashes in URL.
   // Also to avoid name clashes with existing URLs, replace only the first
   // occurrence of the base path.
+  // Also, the 'url-current-[path]' is added so we can distinguish between
+  // 'user' and 'user/login' pages easily and without duplicating CSS.
   // Save some overhead
   $request_uri = request_uri();
   $base_path   = base_path();
@@ -28,21 +30,25 @@ function ultima_preprocess_html(&$vars) {
   switch (count($url_parts)) {
     case 1:
       $vars['classes_array'][] = 'url-' . $url_parts[0];
+      $vars['classes_array'][] = 'url-current-' . $url_parts[0];
       break;
     case 2:
       $vars['classes_array'][] = 'url-' . $url_parts[0];
       $vars['classes_array'][] = 'url-' . $url_parts[0] . '-' . $url_parts[1];
+      $vars['classes_array'][] = 'url-current-' . $url_parts[0] . '-' . $url_parts[1];
       break;
     case 3:
       $vars['classes_array'][] = 'url-' . $url_parts[0];
       $vars['classes_array'][] = 'url-' . $url_parts[0] . '-' . $url_parts[1];
       $vars['classes_array'][] = 'url-' . $url_parts[0] . '-' . $url_parts[1] . '-' . $url_parts[2];
+      $vars['classes_array'][] = 'url-current-' . $url_parts[0] . '-' . $url_parts[1] . '-' . $url_parts[2];
       break;
     case 4:
       $vars['classes_array'][] = 'url-' . $url_parts[0];
       $vars['classes_array'][] = 'url-' . $url_parts[0] . '-' . $url_parts[1];
       $vars['classes_array'][] = 'url-' . $url_parts[0] . '-' . $url_parts[1] . '-' . $url_parts[2];
       $vars['classes_array'][] = 'url-' . $url_parts[0] . '-' . $url_parts[1] . '-' . $url_parts[2] . '-' . $url_parts[3];
+      $vars['classes_array'][] = 'url-current-' . $url_parts[0] . '-' . $url_parts[1] . '-' . $url_parts[2] . '-' . $url_parts[3];
       break;
     case 5:
       $vars['classes_array'][] = 'url-' . $url_parts[0];
@@ -50,6 +56,7 @@ function ultima_preprocess_html(&$vars) {
       $vars['classes_array'][] = 'url-' . $url_parts[0] . '-' . $url_parts[1] . '-' . $url_parts[2];
       $vars['classes_array'][] = 'url-' . $url_parts[0] . '-' . $url_parts[1] . '-' . $url_parts[2] . '-' . $url_parts[3];
       $vars['classes_array'][] = 'url-' . $url_parts[0] . '-' . $url_parts[1] . '-' . $url_parts[2] . '-' . $url_parts[3] . '-' . $url_parts[4];
+      $vars['classes_array'][] = 'url-current-' . $url_parts[0] . '-' . $url_parts[1] . '-' . $url_parts[2] . '-' . $url_parts[3] . '-' . $url_parts[4];
       break;
   }